+++ changelog.md
@@ -18,6 +18,7 @@
 - Projectiles from B. B. Gun, Blackjack, and English can now fall into pits when at rest
 - Increased Blackjack's ammo from 104 to 156 to compensate for new pitfall behavior
 - B. B. Gun's projectiles are now automatically picked up when walking near them while at rest, similar to Blackjack's projectiles
+- Added shaders to mastered guns as an extra indicator of mastery status
 
 ## 1.26.3 (2025-06-22)